requirements.txt
playwright
pytz
python-dotenv
requests
